removed tester from main to avoid bad behavior
