+++ PVGeo/__init__.py
@@ -11,7 +11,7 @@ from . import tunneling
 from . import ubc
 
 from . import _helpers
-from .__main__ import test
+from .__tester__ import test
 
 __author__ = 'Bane Sullivan'
 __license__ = 'BSD-3-Clause'
